+++ packages/react/src/utils/mobx-react-utils.js
@@ -42,3 +42,6 @@ export let lenservable = x => {
 export let withStateLens = state => inject(() => F.lensOf(observable(state)))
 
 export let partial = _.curry((x, y) => inject(() => x)(y))
+
+export let injectDefaults = f =>
+  inject((stores, props) => f(_.defaults(stores, props)))